chore: reorganize utilities, add visuallyHidden util (#185)
* chore: rename hasWindow and update its logic
