+++ src/shared/utilities/canUseDom.test.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Enzyme from 'enzyme';
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import { canUseDom } from './canUseDom';
+
+Enzyme.configure({ adapter: new Adapter() });
+
+describe('canUseDom', () => {
+    test('should return true when the function is called in the browser', () => {
+        expect(canUseDom()).toBe(true);
+    });
+});
