+++ src/shared/utilities/debounce.ts
@@ -1,0 +1,27 @@
+/**
+ * https://dev.to/bwca/create-a-debounce-function-from-scratch-in-typescript-560m
+ * @param fn Accepted argument and return, a function.
+ * @param ms The delay time.
+ * @returns a debounced function that resolves to a value after certain amount of time. Or never resolves if terminated.
+ */
+export function debounce<A = unknown, R = void>(
+    fn: (args: A) => R,
+    ms: number
+): [(args: A) => Promise<R>, () => void] {
+    let timer: NodeJS.Timeout;
+
+    const debouncedFunc = (args: A): Promise<R> =>
+        new Promise((resolve) => {
+            if (timer) {
+                clearTimeout(timer);
+            }
+
+            timer = setTimeout(() => {
+                resolve(fn(args));
+            }, ms);
+        });
+
+    const teardown = () => clearTimeout(timer);
+
+    return [debouncedFunc, teardown];
+}
