+++ src/shared/utilities/generateId.ts
@@ -1,0 +1,6 @@
+/**
+ * Get unique id
+ */
+export const generateId = (prefix?: string) => {
+    return `${prefix ?? ''}${Math.random().toString(36).substring(2, 9)}`;
+};
